+++ backend/src/user/dto/ban-user.dto.ts
@@ -6,9 +6,5 @@ export class BanUserDto {
     example: "You have been banned for bad words",
     description: "ban a user for violating the rules ",
   })
-  @IsString({ message: "Must be a string" })
-  @Length(6, 40, {
-    message: "Ban reason must be longer then 6 symbols and smaller then 20 symbols",
-  })
   readonly banReason: string;
 }